+++ src/components/button/button.component.spec.ts
@@ -21,7 +21,7 @@ describe('Button:', () => {
     it('should bind the "disabled" property',
         async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) =>
             tcb.overrideTemplate(TestComponent, `
-                <gtx-button [disabled]='true'></gtx-button>
+                <gtx-button [disabled]="true"></gtx-button>
             `)
             .createAsync(TestComponent)
             .then(fixture => {
@@ -36,7 +36,7 @@ describe('Button:', () => {
     it('should accept literal "disabled" property',
         async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) =>
             tcb.overrideTemplate(TestComponent, `
-                <gtx-button disabled='true'></gtx-button>
+                <gtx-button disabled="true"></gtx-button>
             `)
             .createAsync(TestComponent)
             .then(fixture => {
@@ -48,6 +48,68 @@ describe('Button:', () => {
         ))
     );
 
+    it('should accept empty "disabled" property',
+        async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) =>
+            tcb.overrideTemplate(TestComponent, `
+                <gtx-button disabled></gtx-button>
+            `)
+            .createAsync(TestComponent)
+            .then(fixture => {
+                let button: HTMLButtonElement = fixture.nativeElement.querySelector('button');
+                fixture.detectChanges();
+                expect(button.disabled).toBe(true);
+            })
+        ))
+    );
+
+    it('forwards its "click" event when enabled',
+        async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) =>
+            tcb.overrideTemplate(TestComponent, `
+                <gtx-button (click)="onClick($event)"></gtx-button>
+            `)
+            .createAsync(TestComponent)
+            .then(fixture => {
+                let onClick = fixture.componentRef.instance.onClick = jasmine.createSpy('onClick');
+                let button: HTMLButtonElement = fixture.nativeElement.querySelector('button');
+                fixture.detectChanges();
+
+                let event = document.createEvent('MouseEvent');
+                event.initEvent('click', true, true);
+
+                button.dispatchEvent(event);
+                button.parentElement.dispatchEvent(event);
+                button.click();
+
+                expect(onClick).toHaveBeenCalledTimes(3);
+            })
+        ))
+    );
+
+    // Disabled elements don't fire mouse events in some browsers, but not all
+    // http://stackoverflow.com/a/3100395/5460631
+    it('does not forward "click" event when disabled',
+        async(inject([TestComponentBuilder], (tcb: TestComponentBuilder) =>
+            tcb.overrideTemplate(TestComponent, `
+                <gtx-button [disabled]="true" (click)="onClick($event)"></gtx-button>
+            `)
+            .createAsync(TestComponent)
+            .then(fixture => {
+                let onClick = fixture.componentRef.instance.onClick = jasmine.createSpy('onClick');
+                let button: HTMLButtonElement = fixture.nativeElement.querySelector('button');
+                fixture.detectChanges();
+
+                let event = document.createEvent('MouseEvent');
+                event.initEvent('click', true, true);
+
+                button.dispatchEvent(event);
+                button.parentElement.dispatchEvent(event);
+                button.click();
+
+                expect(onClick).not.toHaveBeenCalled();
+            })
+        ))
+    );
+
 });
 
 @Component({
@@ -55,4 +117,5 @@ describe('Button:', () => {
     directives: [Button]
 })
 class TestComponent {
+    onClick(): void {}
 }